LayoutTests/media/video-currenttime-monotonic-expected.txt

Test currentTime always monotonically increase.


EVENT(seeked)
EXPECTED (video.currentTime >= '5.7272') OK
EVENT(ended)
EXPECTED (6.0272 == '6.0272') OK
EXPECTED (video.currentTime <= '0.5') OK
END OF TEST
